Version read from VERSION file if not in git
If the package is build from outside git repository (e.g after
exporting source tarball) and git repository is not available
the versio is taken from VERSION file instead of git describe
command. If VERSION file is also unavailable then the version is "?"

Also removed dangerous and no longer necessary make-srcpkg scipt.
Crating source tarballs is not necessary as git pull is to be used
to retrieve icd source code.